More Comments
Made better comments for the class methods, and at the top of the files.
There shouldn't be any more code commits for a while, if I ever decide
to improve the game.
List of things to improve -
1) Better feedback on winning/losing. A buzzer, better flashing, not
sure yet.
2) Make the Score/Leaderboard classes keep track of how well you do.
Right now there's only a way to lose.
3) Fix the button debouncing. The buttons are pretty good, but bounce
every now and then. Haven't decided on hardware/software implementation
of debouncing yet.